production_change: Add stockpile averaging
Production changes based on stockpile level tend to jump between the
most extreme limits, resulting in production capacity that fluctuates
between too much and too little by a large degree.

This is due to the monthly production callback which takes the current
stockpile level which may be just before or after a delivery. In either
case it does not accurately reflect how much input is available.

This commit introduces a simple moving average (SMA) that tracks the
stockpile level, smoothing out these changes.