prod_change: Remove FIXME for randomRange
The idea was that one might want to change the maximum production rate,
and so the randomRange() parameters would need to change. However given
that NML doesn't really deal with floating point numbers it would be
very difficult to compute a range that gives a good percentage change
over the full range of production rates.

So, I will stick with the values from the OpenTTD source code.